Remove EventInputProcessor (again fold input into GameEngine module)
