+++ core/src/com/fallenflame/game/GDXRoot.java
@@ -27,8 +27,6 @@ public class GDXRoot extends Game implements ScreenListener {
 	private LoadingMode loading;
 	/** Player mode for the the game */
 	private GameEngine engine;
-	/**Input Processor for the game itself */
-	private EventInputProcessor eventInputProcessor;
 
 	/**
 	 * Creates a new game from the configuration settings.
@@ -45,12 +43,11 @@ public class GDXRoot extends Game implements ScreenListener {
 		canvas  = new GameCanvas();
 		loading = new LoadingMode(canvas,1);
 		engine = new GameEngine();
-		eventInputProcessor = new EventInputProcessor(engine);
 		InputMultiplexer multiplexer = new InputMultiplexer(); //Allows for multiple InputProcessors
 		//Multiplexer is an ordered list, so when an event occurs, it'll check loadingMode first, and then
 		// LightInputProcessor
 		multiplexer.addProcessor(loading);
-		multiplexer.addProcessor(eventInputProcessor);
+		multiplexer.addProcessor(engine);
 		Gdx.input.setInputProcessor(multiplexer);
 
 		// Initialize the three game worlds
