+++ core/src/com/fallenflame/game/GameEngine.java
@@ -2,6 +2,7 @@ package com.fallenflame.game;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
+import com.badlogic.gdx.InputProcessor;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
@@ -17,7 +18,7 @@ import com.fallenflame.game.util.ScreenListener;
  * Credit to @author: Professor White for code used in this class
  */
 
-public class GameEngine implements Screen {
+public class GameEngine implements Screen, InputProcessor {
     /**Enum to determine if we are loading or not */
     protected enum AssetState {
         /** No assets loaded */
@@ -343,13 +344,6 @@ public class GameEngine implements Screen {
         prevFailed = isFailed;
     }
 
-    /**Increments the player's light radius
-     * @param amount, which represents the amount to increment the light radius*/
-
-    public void lightFromPlayer(float amount){
-        level.lightFromPlayer(amount);
-    }
-
     /**
      * Draw the physics objects to the canvas
      *
@@ -458,6 +452,8 @@ public class GameEngine implements Screen {
         this.listener = listener;
     }
 
+    /************************ POLLING INPUT HANDLING ************************/
+
     /**
      * Get mouse position
      * @return Vector2 mouse position
@@ -527,6 +523,83 @@ public class GameEngine implements Screen {
         if (Gdx.input.isKeyPressed(Input.Keys.S)) {
             vertical -= 1.0f;
         }
+    }
+
+    /************************ EVENT-BASED INPUT HANDLING ************************/
+
+    /**Below are functions that are required to be implemented by InputProcessor. All return false if unused to indicate
+     * that the event was not handled by this inputProcessor. The functions that return false in this file return true
+     * in the LoadingMode inputProcessor to ensure that the even is handled (either LoadingMode does something
+     * or just returns true). This has to be used because mouse scrolling can only be done with InputProcessor.
+     * Gdx.Input does not have any functions to handle mouse scrolling and this should not take too much time,
+     * even with the events*/
+
+    /**What happens when a key is pressed
+     * @param keycode representing what key was pressed
+     * @return boolean saying if the event was handled*/
+    public boolean keyDown (int keycode) {
+        return false;
+    }
+
+    /**What happens when a key is released
+     * @param keycode representing what key was released
+     * @return boolean saying if the event was handled*/
+    public boolean keyUp (int keycode) {
+        return false;
+    }
+
+
+    /**What happens when a character is typed
+     * @param character representing what character was typed
+     * @return boolean saying if the event was handled*/
+    public boolean keyTyped (char character) {
+        return false;
+    }
+
+
+    /**What happens when a screen is touched (for phone) or mouse
+     * @param pointer, button representing what where and what was touched
+     * @return boolean saying if the event was handled*/
+    public boolean touchDown (int x, int y, int pointer, int button) {
+        return false;
+    }
+
+    /**What happens when a screen is released (for phone) or mouse
+     * @param pointer, button representing what where and what was released
+     * @return boolean saying if the event was handled*/
+    public boolean touchUp (int x, int y, int pointer, int button) {
+        return false;
+    }
+
+    /**What happens when a screen is dragged with a finger (for phone) or mouse
+     * @param pointer, button representing what where and what was dragged
+     * @return boolean saying if the event was handled*/
+    public boolean touchDragged (int x, int y, int pointer) {
+        return false;
+    }
+
+    /**What happens when the mouse is moved
+     * @param x, y representing where the mouse moved
+     * @return boolean saying if the event was handled*/
+    public boolean mouseMoved (int x, int y) {
+        return false;
+    }
+
+    /**What happens when the mouse is scrolling. Should take O(1).
+     * @param amount representing if the wheel scrolled down (1) or up (-1). Can only be those two values.
+     * @return boolean saying if the event was handled*/
+    public boolean scrolled (int amount) {
+        if(!isScreenActive()){
+            return true;
+        }
+        if(amount == 1){
+            level.lightFromPlayer(-1.0f);
+        }
+        if(amount == -1){
+            level.lightFromPlayer(1.0f);
+        }
+
+        return true;
     }
 
 }
